Add creation test for Hashtag
